cast adc value to float

+++ firmware/conversion.c
@@ -14,6 +14,6 @@ float temperature_adc(void) {
   itoa( adc, buffer, 10);   // convert interger into string (decimal format)         
   uart_puts(buffer);        // and transmit string to UART
   uart_puts_P("\r\n");
-  return ( TEMP_MIN+(TEMP_MAX-TEMP_MIN)/(ADC_MAX-ADC_MIN) * adc );
+  return ( TEMP_MIN+(TEMP_MAX-TEMP_MIN)/(ADC_MAX-ADC_MIN) * (float)adc );
 }
 #endif /* CONVERSION_H */
